+++ src/dplyr.cpp
@@ -89,7 +89,10 @@ Result* count_distinct_prototype(SEXP call, const LazySubsets& subsets, int){
 }
 
 Result* row_number_prototype(SEXP call, const LazySubsets& subsets, int nargs ){
-    if( nargs != 1) return 0;
+    if( nargs >  1 ) return 0;
+    
+    if( nargs == 0 ) return new RowNumber_0() ;
+    
     Armor<SEXP> data( CADR(call) );
     if( TYPEOF(data) == LANGSXP && CAR(data) == Rf_install("desc") ){
         data = CADR(data) ;